Rename modules whose names conflict with stdlib
